moved the WorldFeature related util to util in WorldFeature directory
