+++ src/main/java/teamport/aether/world/feature/dungeon/silver/component/WorldFeatureSilverMaze.java
@@ -5,14 +5,14 @@ import net.minecraft.core.util.helper.Direction;
 import net.minecraft.core.util.helper.DyeColor;
 import teamport.aether.AetherMod;
 import teamport.aether.blocks.AetherBlocks;
-import teamport.aether.helper.MazeHelper;
+import teamport.aether.world.feature.util.MazeHelper;
 import teamport.aether.helper.unboxed.IntPair;
 import teamport.aether.world.feature.util.BlockPallet;
 import teamport.aether.world.feature.util.WorldFeatureComponent;
 
 import java.util.*;
 
-import static teamport.aether.helper.MetadataHelper.*;
+import static teamport.aether.world.feature.util.MetadataHelper.*;
 import static teamport.aether.world.feature.dungeon.silver.WorldFeatureAetherSilverDungeon.ANGELIC;
 import static teamport.aether.world.feature.util.WorldFeatureBlock.wfb;
 import static teamport.aether.world.feature.util.WorldFeatureComponent.*;
