+++ src/main/java/teamport/aether/world/feature/util/MazeHelper.java
@@ -1,4 +1,4 @@
-package teamport.aether.helper;
+package teamport.aether.world.feature.util;
 
 import teamport.aether.helper.unboxed.IntPair;
